refactor: reorganize project folder structure.
