+++ middlewares/security/loginLimiter.js
@@ -1,6 +1,6 @@
 import rateLimit, {ipKeyGenerator} from "express-rate-limit";
 import { RedisStore } from "rate-limit-redis";
-import { redisClient } from "../../configs/redis.js";
+import { redisClient } from "../../config/redis.js";
 
 export const loginLimiter = rateLimit({
     windowMs: 15 * 60 * 1000, // 15 minutes
